requirements.txt
dj-database-url==0.4.1
Django==1.9.8
django-ordered-model==1.2.1
gunicorn==19.6.0
psycopg2==2.6.2
whitenoise==3.2
